Split out legacy enzyme tests and use `<RenderWithContext>`
